requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==8.1
djangorestframework==3.11
dj-rest-auth==1.0.4
python-decouple==3.3

